refactor(app-board): remove react-zoom-pan-pinch dep and extract function
