+++ packages/game-app/package.json
@@ -14,7 +14,6 @@
     "react-hook-form": "^6.13.1",
     "react-redux": "^7.2.2",
     "react-router-dom": "^5.2.0",
-    "react-zoom-pan-pinch": "^1.6.1",
     "redux-saga": "^1.1.3",
     "styled-components": "^5.2.1",
     "styled-system": "^5.1.5",
