+++ packages/game-app/src/gameView/components/ZoomPanContext/ZoomPanContext.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+type Props = {};
+
+const ZoomPanContext: React.FC<Props> = ({}) => {
+  return <></>;
+};
+
+ZoomPanContext.displayName = 'ZoomPanContext';
+
+export default ZoomPanContext;
