+++ packages/game-app/src/gameView/components/ZoomPanContext/index.ts
@@ -1,0 +1,3 @@
+import ZoomPanContext from './ZoomPanContext';
+
+export default ZoomPanContext;
