+++ packages/game-app/src/temporary.css
@@ -141,6 +141,7 @@ label {
   width: 100vw;
   height: 100vh;
   overflow: hidden;
+  position: relative;
 }
 
 .react-transform-component {
